adjust file & folder structure
