+++ deskmark/app/components/CreateBar/index.jsx
@@ -1,6 +1,8 @@
 /*
- * @file component create-bar
+ * @file component CreateBar
  */
+
+import './style.scss';
 
 import React, {PropTypes} from 'react';
 
@@ -10,7 +12,7 @@ const propTypes = {
 
 function CreateBar({ onClick }){
   return (
-    <a href="#" onClick={onClick} className="list-group-item create-entry">
+    <a href="#" onClick={onClick} className="list-group-item create-bar-component">
       + 创建新的文章
     </a>
   );
