+++ deskmark/app/components/CreateBar/style.scss
@@ -1,0 +1,13 @@
+/*
+ * @file style for component CreateBar
+ */
+
+a.create-bar-component {
+  background: #0275d8;
+  color: #fff;
+
+  &:hover {
+    background: darken(#0275d8, 20%);
+    color: #fff;
+  }
+}
