+++ deskmark/app/components/Deskmark/style.scss
@@ -1,0 +1,9 @@
+/*
+ * @file style for component Deskmark
+ */
+
+.deskmark-component {
+  .container {
+    margin-top: 50px;
+  }
+}
