+++ deskmark/app/components/ItemEditor/style.scss
@@ -1,8 +1,8 @@
 /*
- * @file style for component editor
+ * @file style for component ItemEditor
  */
 
-.editor-component {
+.item-editor-component {
   padding: 10px;
   border-top: 1px solid #efefef;
   position: relative;
