+++ deskmark/app/components/ItemShowLayer/style.scss
@@ -1,8 +1,8 @@
 /*
- * @file style for component shower
+ * @file style for component ItemShowLayer
  */
 
-.shower-component {
+.item-show-layer-component {
   padding: 10px;
   border-top: 1px solid #efefef;
   position: relative;
