--- deskmark/app/components/deskmark/deskmark.scss
@@ -1,9 +1,0 @@
-/*
- * @file style for component deskmark
- */
-
-.desk-mark-component {
-  .container {
-    margin-top: 50px;
-  }
-}
