--- deskmark/app/components/list/list.scss
@@ -1,14 +1,0 @@
-/*
- * @file style for component list
- */
-
-.list-component {
-  .create-entry {
-    background: #0275d8;
-    color: #fff;
-    &:hover {
-      background: darken(#0275d8, 20%);
-      color: #fff;
-    }
-  }
-}
